Adds operations configuration support
Adds operations configuration support to the benchflow benchmark
configuration file model; it is now possible to specify the `sut_type`
as either `http` or `wfms`; in the first case, it will be possible to
specify an `http` driver with operations describing http requests, such
as:

```
myOperation:
    method: get
    endpoint: /myOperation
    data: request body
    headers:
         Content-Type: text/plain
```

while in the case of a `wfms` driver it will be possible to refer to
each model, and specify data to pass in the body of the request of the
model; moreover, we provide a type safe representation of the types of
drivers supported for WfMSs; currently, only a `start` driver is
supported.

The logic for the generation of the faban configuration xml file is
updated accordingly to the above changes.

Also, includes minor refactorings (mostly package organization changes).
