+++ application/src/main/scala/cloud/benchflow/benchmark/config/DeploymentDescriptorBuilder.scala
@@ -1,7 +1,8 @@
-package cloud.benchflow.config
+package cloud.benchflow.benchmark.config
 
-import cloud.benchflow.config.benchflowbenchmark.BenchFlowBenchmark
-import cloud.benchflow.config.docker.compose.DockerCompose
+import cloud.benchflow.benchmark.config._
+import cloud.benchflow.benchmark.config.benchflowbenchmark.BenchFlowBenchmark
+import cloud.benchflow.benchmark.config.docker.compose.DockerCompose
 import cloud.benchflow.driversmaker.requests.Trial
 import cloud.benchflow.driversmaker.utils.env.DriversMakerBenchFlowEnv
