+++ application/src/test/resources/benchflow-benchmark.yml
@@ -21,10 +21,14 @@ properties:
         maxRunTime: "6"
         interval: "30"
 drivers:
-    - cloud.benchflow.wfmsbenchmark.driver.WfMSBenchmarkDriver:
-        agents: 10.7.0.10:1
-        stats:
-            interval: "30"
+    - start:
+        properties:
+            agents: 10.7.0.10:1
+            stats:
+                interval: "30"
+        operations:
+            - myModel:
+                 data: ciaociao
 sut-configuration:
     target-service:
         name: camunda
